backend/requirements.txt
# asgiref==3.4.1
# certifi==2021.10.8
# Django==3.2.10
# djangorestframework==3.13.1
# pytz==2021.3
# sqlparse==0.4.2
# typing_extensions==4.0.1
# Pillow== 8.1
# asgiref==3.4.1
# certifi==2021.10.8
# dj-database-url==0.5.0
# Django==3.2.10
# #django-cors-headers==3.10.1
# #django-mongodb-engine-py3==0.6.0.1
# #django-rest-framework-mongoengine==3.4.1
# djangorestframework==3.13.1
# #djangotoolbox==1.8.0
# djongo==1.3.6
# dnspython==2.1.0
# mongoengine==0.23.1
# Pillow==8.1.0
# pymongo==3.12.3
# pytz==2021.3
# sqlparse==0.2.4
# typing_extensions==4.0.1

asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.10
django-cors-headers==3.10.1
django-htmlmin==0.11.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djongo==1.3.6
drf-yasg==1.20.0
flex==6.14.1
html5lib==1.1
idna==3.3
importlib-metadata==4.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonpointer==2.2
MarkupSafe==2.0.1
packaging==21.3
Pillow==9.0.0
PyJWT==1.7.1
pymongo==3.12.3
pyparsing==3.0.6
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rfc3987==1.3.8
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.1
sqlparse==0.2.4
strict-rfc3339==0.7
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
validate-email==1.3
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==5.3.0
<<<<<<< HEAD
zipp==3.7.0
=======
zipp==3.7.0
celery
>>>>>>> docker
